# static/swagger.spec.json
{
  "info": {
    "title": "gearlaunch2-backend-rest-demo",
    "version": "1.0.0",
    "description": "GearLaunch v2 Rest Demo Backend"
  },
  "host": "localhost:8000",
  "schemes": [
    "http",
    "https"
  ],
  "basePath": "/",
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Standard Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "swagger": "2.0",
  "paths": {
    "/api/accounts/v1/countries": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Find all Countries",
        "description": "This API returns a list of all `Countries` based on common `select` request.\n",
        "operationId": "findAllCountries",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "select",
            "in": "query",
            "type": "object",
            "description": "The `select` query object is used to point what data the API need to return.\nAdding a property to the `select` object will return the corresponding data.\n\nFor example, if you'd like to get a list of all `Countries` with their continent, you can submit this `select`:\n```\n{\n  \"name\": \"\",\n  \"continent_name\": \"\"\n}\n```\n\nFiltration is also supported with the special key of `\"$where\":{}`. The `$where` object supports all comparison operators\nfrom Sequelize ORM https://sequelize.org/master/manual/model-querying-basics.html#operators. Please note that operations are prefixed, e.g. instead of `Op.eq` => `$eq`.\n\nFor example, if you'd like to get a list of all `Countries` within the continent of `Africa`, you can submit this `select`:\n```\n{\n  \"name\": \"\",\n  \"$where\": {\n    \"continent_name\": {\n      \"$eq\": \"Africa\"\n    }\n  }\n}\n```\n",
            "default": {
              "code": "",
              "name": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of `Countries` that matched the request parameters.\n",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DatabaseCountry"
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      }
    },
    "/api/accounts/v1/merchants": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "summary": "Find all Merchants",
        "description": "This API returns a list of all `Merchants` based on common `select` request.\n",
        "operationId": "findAllMerchants",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "select",
            "in": "query",
            "type": "object",
            "description": "The `select` query object is used to point what data the API need to return.\nAdding a property to the `select` object will return the corresponding data.\n\nFor example, if you'd like to get a list of all `Merchants`, you can submit this `select`:\n```\n{\n  \"merchant_name\": \"\"\n}\n```\n\nFiltration is also supported with the special key of `\"$where\":{}`. The `$where` object supports all comparison operators\nfrom Sequelize ORM https://sequelize.org/master/manual/model-querying-basics.html#operators. Please note that operations are prefixed, e.g. instead of `Op.eq` => `$eq`.\n\nFor example, if you'd like to get a list of all `Merchants` within the continent of `Africa`, you can submit this `select`:\n```\n{\n  \"merchant_name\": \"\",\n  \"country\": {\n    \"name\": \"\",\n    \"continent_name\": \"\",\n    \"$where\": {\n      \"continent_name\": {\n        \"$eq\": \"Africa\"\n      }\n    }\n  }\n}\n```\n",
            "default": {
              "merchant_name": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of `Merchants` that matched the request parameters.\n",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DatabaseMerchant"
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      }
    },
    "/api/products/v1/order_items": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Find all OrderItems",
        "description": "This API returns a list of all `OrderItems` based on common `select` request.\n",
        "operationId": "findAllOrderItems",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "select",
            "in": "query",
            "type": "object",
            "description": "The `select` query object is used to point what data the API need to return.\nAdding a property to the `select` object will return the corresponding data.\n\nFor example, if you'd like to get a list of all `OrderItems`, you can submit this `select`:\n```\n{\n  \"quantity\": \"\"\n}\n```\n\nFiltration is also supported with the special key of `\"$where\":{}`. The `$where` object supports all comparison operators\nfrom Sequelize ORM https://sequelize.org/master/manual/model-querying-basics.html#operators. Please note that operations are prefixed, e.g. instead of `Op.eq` => `$eq`.\n\nFor example, if you'd like to get a list of all `OrderItems` that are `in_progress` for all buyer `Users` in Africa, you can submit this `select`:\n```\n{\n  \"quantity\": \"\",\n  \"product\": {\n    \"name\": \"\"\n  },\n  \"order\": {\n    \"status\": \"\",\n    \"$where\": {\n      \"status\": {\n        \"$eq\": \"in_progress\"\n      }\n    },\n    \"user\": {\n      \"full_name\": \"\",\n      \"country\": {\n        \"name\": \"\",\n        \"continent_name\": \"\",\n        \"$where\": {\n          \"continent_name\": {\n            \"$eq\": \"Africa\"\n          }\n        }\n      }\n    }\n  }\n}\n```\n",
            "default": {
              "quantity": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of `OrderItems` that matched the request parameters.\n",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DatabaseOrderItem"
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      }
    },
    "/api/products/v1/orders": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Find all Orders",
        "description": "This API returns a list of all `Orders` based on common `select` request.\n",
        "operationId": "findAllOrders",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "select",
            "in": "query",
            "type": "object",
            "description": "The `select` query object is used to point what data the API need to return.\nAdding a property to the `select` object will return the corresponding data.\n\nFor example, if you'd like to get a list of all `Orders`, you can submit this `select`:\n```\n{\n  \"status\": \"\"\n}\n```\n\nFiltration is also supported with the special key of `\"$where\":{}`. The `$where` object supports all comparison operators\nfrom Sequelize ORM https://sequelize.org/master/manual/model-querying-basics.html#operators. Please note that operations are prefixed, e.g. instead of `Op.eq` => `$eq`.\n\nFor example, if you'd like to get a list of all `Orders` with status `in_progress` that were bought by `Users` from Africa, you can submit this `select`:\n```\n{\n  \"status\": \"\",\n  \"$where\": {\n    \"status\": {\n      \"$eq\": \"in_progress\"\n    }\n  },\n  \"user\": {\n    \"full_name\": \"\",\n    \"country\": {\n      \"name\": \"\",\n      \"continent_name\": \"\",\n      \"$where\": {\n        \"continent_name\": {\n          \"$eq\": \"Africa\"\n        }\n      }\n    }\n  }\n}\n```\n",
            "default": {
              "status": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An array of `Orders` that matched the request parameters.\n",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DatabaseOrder"
              }
            }
          },
          "500": {
            "$ref": "#/responses/AppNotFound"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ResponseError": {
      "type": "object",
      "description": "Schema for `4XX` server error response body.\nResponse includes a UUIDv4, error code and message body.\n",
      "properties": {
        "errorId": {
          "type": "string",
          "format": "uuid",
          "example": "a34f4558-1b88-47cd-8d63-ae4eaf038bd7"
        },
        "errorCode": {
          "type": "string",
          "example": "GENERIC_SERVER_ERROR"
        },
        "errorMessage": {
          "type": "string",
          "example": "Generic server error"
        }
      },
      "required": [
        "errorId",
        "errorCode",
        "errorMessage"
      ]
    },
    "DatabaseCountry": {
      "description": "List of all `countries` that can be related to other entities such as `users` and `merchants`.\n",
      "allOf": [
        {
          "type": "object",
          "required": [
            "code",
            "name",
            "continent_name"
          ],
          "properties": {
            "code": {
              "type": "number",
              "description": "Unique country ID",
              "example": 34
            },
            "name": {
              "type": "string",
              "description": "country name",
              "example": "Vanuatu"
            },
            "continent_name": {
              "type": "string",
              "description": "On which continent this country is located",
              "example": "Africa"
            }
          }
        }
      ]
    },
    "DatabaseMerchant": {
      "description": "`Merchants` are users who can sell `products`.\n",
      "allOf": [
        {
          "type": "object",
          "required": [
            "id",
            "merchant_name",
            "admin_id",
            "country_code",
            "created_at"
          ],
          "properties": {
            "id": {
              "type": "number",
              "description": "Unique merchant ID",
              "example": 34
            },
            "merchant_name": {
              "type": "string",
              "description": "How the user would like to be named as merchant",
              "example": "Jamie Stone"
            },
            "admin_id": {
              "type": "number",
              "description": "Association to the `users` entity. Shows which user uses this merchant entity.",
              "example": 13
            },
            "country_code": {
              "type": "number",
              "description": "Association to the country entity. Shows user country address.",
              "example": 22
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "readOnly": true,
              "description": "DateTime indicating when that merchant was created",
              "example": "2020-04-30T00:00:00.000Z"
            }
          }
        }
      ]
    },
    "DatabaseOrderItem": {
      "description": "`Order` items are the collection of all bought `products` in a specific `order`.\n",
      "allOf": [
        {
          "type": "object",
          "required": [
            "id",
            "product_id",
            "order_id",
            "quantity"
          ],
          "properties": {
            "id": {
              "type": "number",
              "description": "Unique order item ID",
              "example": 34
            },
            "product_id": {
              "type": "number",
              "description": "Association to the `products` entity. Shows which product was bought.",
              "example": 100
            },
            "order_id": {
              "type": "number",
              "description": "Association to the `orders` entity. Shows which order bought the related product.",
              "example": 200
            },
            "quantity": {
              "type": "number",
              "description": "Shows how much of the product was bought.",
              "example": 5
            }
          }
        }
      ]
    },
    "DatabaseOrder": {
      "description": "`Orders` are collection of bought `order items`.\n",
      "allOf": [
        {
          "type": "object",
          "required": [
            "id",
            "user_id",
            "status",
            "created_at"
          ],
          "properties": {
            "id": {
              "type": "number",
              "description": "Unique order ID",
              "example": 34
            },
            "user_id": {
              "type": "number",
              "description": "Association to the `users` entity. Shows which user bought this product.",
              "example": 14
            },
            "status": {
              "type": "string",
              "description": "A free text field to indicate if the order is paid.",
              "example": "in_progress"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "readOnly": true,
              "description": "DateTime indicating when that order was created",
              "example": "2020-04-30T00:00:00.000Z"
            }
          }
        }
      ]
    },
...
  },
  "responses": {
...
    "AppNotFound": {
      "description": "Some unexpected issue occurred and crashed the app.\n",
      "schema": {
        "$ref": "#/definitions/ResponseError"
      },
      "examples": {
        "application/json": {
          "errorId": "a34f4558-1b88-47cd-8d63-ae4eaf038bd7",
          "errorCode": "INTERNAL_SERVER_ERROR",
          "errorMessage": "Internal server error"
        }
      }
    }
  },
...
}